requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.3
flask-socketio==5.3.6
pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
schedule==1.2.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
requests==2.31.0
python-dateutil==2.8.2
eventlet==0.33.3
werkzeug==2.3.7
# MongoDB dependencies
pymongo==4.6.0
flask-pymongo==2.3.0
mongoengine==0.27.0
bson==0.5.10